requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
black

numpy
pandas
joblib
ta
tqdm
scikit-learn
l3wrapper
torch
pytorch-lightning
comet-ml
imbalanced-learn
seaborn